Made a fork of D3 with no global or jsdom
For easy to use in a browser based module system I removed the global
that is always created so now it can be used form modules with
pollution. Also removed jsdom dependency as it needs contextify, which
is a binary module, which means that you need to install build tools in
windows to properly compile this.